Travel Salesman Problem

+++ src/main/scala/yaol/generic/OptimalDouble.scala
@@ -4,4 +4,4 @@
 package yaol.generic
 
 case class OptimalDouble(val vector: Array[Double], val cost: Double)
-case class OptimalInt(val vector: Array[Int], val cost: Int)
+case class OptimalInt(val vector: Array[Int], val cost: Double)
